swissify_web/requirements.txt
fastapi==0.120.2
uvicorn==0.38.0
jinja2==3.1.6
python-multipart==0.0.20
httpx==0.28.1
python-docx==1.2.0
pypdf==6.1.3
